Fixed: Reverted changes to backend - projects - models, serializers, urls, updated static_analysis - added urls and serializers, upddated models.py and included doc comments.
