+++ projects/serializers.py
@@ -1,13 +1,15 @@
 from rest_framework import serializers
-from .models import AnalysisResult
+from .models import Project
 
-class AnalysisResultSerializer(serializers.ModelSerializer):
-    """
-    Serializer for the AnalysisResult model.
+class ProjectSerializer(serializers.ModelSerializer):
+    #owner_username = serializers.SerializerMethodField()
+    #optional field to display the owner's username, custom method custom serialization logic
+    users = serializers.PrimaryKeyRelatedField(many=True, read_only=True)
 
-    This serializer handles the conversion between AnalysisResult instances and their JSON representations.
-    """
     class Meta:
-        model = AnalysisResult
-        fields = ['id', 'project', 'timestamp', 'status', 'result']
-        read_only_fields = ['id', 'project', 'timestamp', 'status', 'result']
+        model = Project
+        fields = ['id', 'name', 'description', 'repository_url', 'owner', 'owner_username', 'created_at', 'updated_at', 'users']
+        read_only_fields = ['id', 'owner', 'created_at', 'updated_at']
+
+    #def get_owner_username(self, obj):
+    #    return obj.owner.username
