+++ projects/urls.py
@@ -1,7 +1,9 @@
 from django.urls import path
-from .views import InitiateAnalysisView, RetrieveAnalysisResultsView
+from .views import ProjectListCreateView, ProjectRetrieveUpdateDestroyView, UserProjectListView, ProjectUserAddView
 
 urlpatterns = [
-    path('projects/<int:pk>/analyze/', InitiateAnalysisView.as_view(), name='initiate-analysis'),
-    path('projects/<int:pk>/analysis-results/', RetrieveAnalysisResultsView.as_view(), name='analysis-results'),
+    path('projects/', ProjectListCreateView.as_view(), name='project-list-create'),
+    path('projects/<int:pk>/', ProjectRetrieveUpdateDestroyView.as_view(), name='project-detail'),
+    path('users/<int:user_id>/projects/', UserProjectListView.as_view(), name='user-projects'),
+    path('projects/<int:pk>/users/', ProjectUserAddView.as_view(), name='project-add-user'),
 ]
